添加push命令
