+++ cmd/format.go
@@ -2,7 +2,6 @@ package cmd
 
 import (
 	"fmt"
-	"github.com/isxcode/isx-cli/git"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"os"
@@ -40,53 +39,5 @@ func formatCmdMain() {
 		} else {
 			fmt.Println("执行成功")
 		}
-	}
-
-	// 获取当前分支
-	branchName := git.GetCurrentBranchName(viper.GetString("current-project.name"), true)
-
-	// 自动commit 和 提交代码
-	commitAndPushCode(projectPath, branchName)
-
-	// 遍历子模块
-	var subRepository []Repository
-	viper.UnmarshalKey(viper.GetString("current-project.name")+".sub-repository", &subRepository)
-	for _, repository := range subRepository {
-		commitAndPushCode(projectPath+"/"+repository.Name, branchName)
-	}
-
-}
-
-func commitAndPushCode(path string, branchName string) {
-
-	gitAddCommand := "git add ."
-	gitAddCmd := exec.Command("bash", "-c", gitAddCommand)
-	gitAddCmd.Stdout = os.Stdout
-	gitAddCmd.Stderr = os.Stderr
-	gitAddCmd.Dir = path
-	err := gitAddCmd.Run()
-	if err != nil {
-		fmt.Println("git add . 异常")
-	}
-
-	gitCommitCommand := "git commit -m '格式化代码'"
-	gitCommitCmd := exec.Command("bash", "-c", gitCommitCommand)
-	gitCommitCmd.Stdout = os.Stdout
-	gitCommitCmd.Stderr = os.Stderr
-	gitCommitCmd.Dir = path
-	err = gitCommitCmd.Run()
-	if err != nil {
-		fmt.Println("无代码需要commit")
-	}
-
-	// 推送代码
-	pushOriginCommand := "git push origin " + branchName
-	pushOriginCmd := exec.Command("bash", "-c", pushOriginCommand)
-	pushOriginCmd.Stdout = os.Stdout
-	pushOriginCmd.Stderr = os.Stderr
-	pushOriginCmd.Dir = path
-	err = pushOriginCmd.Run()
-	if err != nil {
-		fmt.Println("无法推送，请谨慎尝试强推： isx git push origin " + branchName + " -f")
 	}
 }
